Debugged post restructure
