+++ controllers/admin/book/chapter/readController.js
@@ -1,4 +1,4 @@
-const { Book } = require('../../../../../models');
+const { Book } = require('../../../../models');
 
 module.exports = {
     getChapters: (req, res) => {
